perf hists browser: Fix handling of TAB/UNTAB for multiple events
Patchwork-id: 43507
O-Subject: [PATCH RHEL6.3 BZ784888 44/97] perf hists browser: Fix handling of
	TAB/UNTAB for multiple events
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit 18eaf0b8e60a2fa54667b2192197970174f1c061
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Thu Oct 13 12:22:28 2011 -0300

description
===========
When using multiple events the 'top' and 'report' tools will first
present the user with a menu to choose the event to browse.

After that the user can either press <- to go back to the menu and
choose another event or instead press TAB to go the next event without
having to go back to the menu or shift-TAB (UNTAB) to go the previous
event, useful to quickly visually see if multiple events are correlated.

The handling of each hists browser return was broken by the ed7e566,
that combined both switches, the first that was for choosing the event
and the second that was for checking if switching to the next event
without passing thru the events menu.

Repeat with me: Don't be clever like that.

Fix it by moving the switch to right after the call to the hists
browser, making abundantly clear that the two switches are unrelated.

This also fixes a compiler warning about the 'pos' variable being
possibly used unitialized.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>