perf ui browser: Make the colors configurable and change the defaults
Patchwork-id: 43540
O-Subject: [PATCH RHEL6.3 BZ784888 58/97] perf ui browser: Make the colors
	configurable and change the defaults
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit e039fc727c40c5541bb22aed7f7030dea3e6fb7a
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Tue Oct 18 15:50:51 2011 -0200

description
===========
Just use as a starting point the "[colors]" section of
tools/perf/Documentation/perfconfig.example.

Changed the colors to be the ones in the old perf tool if used in a green on
black xterm.

The next patches should allow using the colors configured for the xterm.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>