perf tools: handle endianness of feature bitmap
Patchwork-id: 43548
O-Subject: [PATCH RHEL6.3 BZ784888 65/97] perf tools: handle endianness of
	feature bitmap
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit d327fa435996c3e559b55e254bad6ed801a92f89
Author: David Ahern <dsahern@gmail.com>
Date:   Tue Oct 18 17:34:01 2011 -0600

description
===========
Feature bitmap is declared as an array of unsigned longs -- not good
since its size can differ between the host that generated the data file
and the host analyzing the file.

We need to handle endianness, but we don't know the size of the unsigned
long where the file was generated. Take a best guess at determining it:
try 64-bit swap first (ie., file created on a 64-bit host), and check if
the hostname feature bit is set (this feature bit is forced on as of
fbe96f2).  If the bit is not, undo the 64-bit swap and try a 32-bit
swap. If the hostname bit is still not set (e.g., older data file), punt
and fallback to the original behavior -- clearing all feature bits and
setting buildid.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>