perf ui: Improve handling sigwinch a bit
Patchwork-id: 43562
O-Subject: [PATCH RHEL6.3 BZ784888 78/97] perf ui: Improve handling sigwinch a
	bit
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit 71172ed97cd4cd45c6ae70e594ba351798d11909
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Tue Oct 25 13:45:16 2011 -0200

description
===========
No need to unblock it at each ui__getch() and also allow other users to
check if a resize is needed, or force an refresh of terminal dimensions.

The 'force' one shouldn't be needed, but its in a slow path, so leave it
like that for now, I'll revisit this another day.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>