perf top tui: Don't recalc column widths considering just the first page
Patchwork-id: 43551
O-Subject: [PATCH RHEL6.3 BZ784888 91/97] perf top tui: Don't recalc column
	widths considering just the first page
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit 104268a335d8aeb4293301e26a7bfad964ddc1ca
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Wed Nov 2 12:46:18 2011 -0200

description
===========
It makes sense for the stdio where we can't navigate to the other pages.

On the TUI it breaks as soon as we navigate to other pages that have,
DSOs with longer names than the ones on the first page.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>