perf top: Fix live annotation in the --stdio interface
Patchwork-id: 43553
O-Subject: [PATCH RHEL6.3 BZ784888 92/97] perf top: Fix live annotation in the
	--stdio interface
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit f9e3d4b1a9c86217655997d3ef109b1eaae967bc
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Thu Nov 3 11:31:26 2011 -0200

description
===========
In the old --stdio interface the annotation is done just after one
selects a symbol, while in --tui, now the default when the required libs
are installed, we annotate all symbols with samples so that when
annotation is asked we see what happened recently on that symbol.

To achieve that the --stdio variant checks if the hist_entry being
processed is the one selected by the user via the 's' hotkey. What
happens now that we share the hist_entry abstractions with 'perf report'
is that for minimizing locking contention multiple rb_trees are used,
one for collecting the samples and other to browse/show them after
resorting it by number of samples and decay them, which is done
periodically.

So the simple test in record_precise_ip doesn't work as we move
hist_entries between those rb_trees. To fix it just check that the
underlying struct symbol associated with those hist_entries is the same.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>