perf: Fix parsing of __print_flags() in TP_printk()
Patchwork-id: 43555
O-Subject: [PATCH RHEL6.3 BZ784888 95/97] perf: Fix parsing of __print_flags()
	in TP_printk()
Bugzilla: 784888
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 784888
https://bugzilla.redhat.com/show_bug.cgi?id=784888

upstream status
===============
commit d06c27b22aa66e48e32f03f9387328a9af9b0625
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Fri Nov 4 16:32:25 2011 -0400

description
===========
A update is made to the sched:sched_switch event that adds some
logic to the first parameter of the __print_flags() that shows the
state of tasks. This change cause perf to fail parsing the flags.

A simple fix is needed to have the parser be able to process ops
within the argument.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>