autofs4: fix expire race
Patchwork-id: 43610
O-Subject: [RHEL-6.3 PATCH] autofs4 - fix expire race
Bugzilla: 766818
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Ian Kent <raven@themaw.net>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=766818

This patch has no upstream counterpart because the vfs-scale
patch series eliminated this bug upstream. It is more risky
to attempt a selective backport than to merge this fairly
straight forward correction. We will need to re-visit this
if the vfs-scale series is to be merged into RHEL-6 (although
I believe that isn't planned).

At the bottom of the loop within autofs4_expire_indirect()
the next dentry in the subdirs list is set after taking the
dcache_lock. But the variable, next, is related to the dentry
currently being used and that dentry is dput() just prior to
taking the dcache_lock. If that dentry is about to go away
then next is invalid and next->next will point to itself, now
an invalid dentry causing an exception when it is accessed at
the top of the loop.

The simplest way to avoid this race is to retain the reference
to the dentry being checked until after the next dentry is set.

Signed-off-by: Ian Kent <raven@themaw.net>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>