sleep: Fix race between CPU hotplug and freezer
Patchwork-id: 43628
O-Subject: [RHEL6 PATCH BZ 747224] PM / Sleep: Fix race between CPU hotplug and
	freezer
Bugzilla: 747224
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Backport of linux-2.6 commit 79cfbdfa87e84992d509e6c1648a18e1d7e68c20

The CPU hotplug notifications sent out by the _cpu_up() and _cpu_down()
functions depend on the value of the 'tasks_frozen' argument passed to them
(which indicates whether tasks have been frozen or not).
(Examples for such CPU hotplug notifications: CPU_ONLINE, CPU_ONLINE_FROZEN,
CPU_DEAD, CPU_DEAD_FROZEN).

Thus, it is essential that while the callbacks for those notifications are
running, the state of the system with respect to the tasks being frozen or
not remains unchanged, *throughout that duration*. Hence there is a need for
synchronizing the CPU hotplug code with the freezer subsystem.

Since the freezer is involved only in the Suspend/Hibernate call paths, this
patch hooks the CPU hotplug code to the suspend/hibernate notifiers
PM_[SUSPEND|HIBERNATE]_PREPARE and PM_POST_[SUSPEND|HIBERNATE] to prevent
the race between CPU hotplug and freezer, thus ensuring that CPU hotplug
notifications will always be run with the state of the system really being
what the notifications indicate, _throughout_ their execution time.

I've tested this on two systems and it seems to resolve the problem.  Having
said that I think that this is sort of a bogus and contrived QA test that
failed.  We've been running RHEL5 and RHEL6 for a long time and no one has
hit this panic -- which IMO means that users are smart enough not to
softplug and suspend at the same time.

BZ: 747224
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>