mptsas: ensure NUL-termination of MptCallbacksName elements
Patchwork-id: 43663
O-Subject: [RHEL6.3 PATCH 09/10] mptsas: ensure NUL-termination of
	MptCallbacksName elements
Bugzilla: 735895
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>

8af3d8d5bc75914a01fe41839474902f2c6b79b9 [SCSI] fusion: ensure NUL-termination of MptCallbacksName elements

Using strlcpy instead of memcpy makes the strlen() calls superfluous
and also ensures zero-termination of the copied string.
At the same time get rid of the magic constant 50.

Actually, I can't see why copying the callback name is necessary
instead of simply storing a pointer to it (just like to the callback
function itself), but that goes beyond the scope of this fix.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>