tcp: bind() fix autoselection to share ports
Patchwork-id: 43721
O-Subject: [RHEL6.3 PATCH] tcp: bind() fix autoselection to share ports
Bugzilla: 784671
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Bugzilla: 784671
https://bugzilla.redhat.com/show_bug.cgi?id=784671

Description:
The current code checks for conflicts when the application
requests a specific port.  If there is no conflict, then
the request is granted.

On the other hand, the port autoselection done by the kernel
fails when all ports are bound even when there is a port
with no conflict available.

The fix changes port autoselection to check if there is a
conflict and use it if not.

Upstream:
commit 2b05ad33e1e624e7f08b8676d270dc7725403b7e
commit fddb7b5761f104f034a0e708ece756d9b2eb2cac

Test Build at:
https://brewweb.devel.redhat.com/taskinfo?taskID=4001739

Testing Status:
Reproduced and verified locally.

Proposed Patch:
It's a backport of the two upstream commits mentioned above.

Signed-off-by: Flavio Leitner <fbl@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>