Set PCIE maxpayload for card during hotplug insertion
Patchwork-id: 43722
O-Subject: [RHEL6.3 PATCH 01/12] PCI: Set PCIE maxpayload for card during
	hotplug insertion
Bugzilla: 692531
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

commit e522a7126c7c144a1dd14c6f217ac31e71082b1d
Author: Jordan_Hargrave Dell com <Jordan_Hargrave Dell com>
Date:   Mon May 9 15:24:55 2011 -0500

PCI: Set PCIE maxpayload for card during hotplug insertion

The following patch sets the MaxPayload setting to match the parent
reading when inserting a PCIE card into a hotplug slot.  On our system,
the upstream bridge is set to 256, but when inserting a card, the card
setting defaults to 128.  As soon as I/O is performed to the card it
starts receiving errors since the payload size is too small.

Reviewed-by: Kenji Kaneshige <kaneshige kenji jp fujitsu com>
Signed-off-by: Jordan Hargrave <jordan_hargrave dell com>
Signed-off-by: Jesse Barnes <jbarnes virtuousgeek org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>