console interrupts vs. panic
Patchwork-id: 43744
O-Subject: [RHEL6.3 PATCH 1/5] [s390x]kernel: console interrupts vs. panic
Bugzilla: 768891
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
Panic action is not executed correctly after a panic occurred.

On panic, the panic message is printed to the console, all but
one cpu are stopped and finally the console buffer is flushed.
One of the stopped cpus could just have received an interrupt
for the console device which causes the console flush to wait
indefinitely.

Add code to smp_send_stop() to wait for outstanding interrupts
on the cpus that are to be stopped.

Bugzilla
--------
BZ 768891
https://bugzilla.redhat.com/show_bug.cgi?id=768891

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.3
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=85ac7ca5972d63d23aa5ea75c3834a33b951f89d

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>