dasd: fix expiration handling for recovery requests
Patchwork-id: 43746
O-Subject: [RHEL6.3 PATCH 2/5] [s390x]dasd: fix expiration handling for recovery
	requests
Bugzilla: 768890
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
A request on a DASD device is blocked and does never finish.

The 'expires' value of a ccw requests defines how long the
device driver should wait for a response from the evice after
the request has been submitted to the channel subsystem. After
the expiration time (e.g. 30 seconds) the waiting request will
be cancelled and started again. This protects the DASD devices
from beeing blocked by errors that cause the answering I/O
interrupt to be lost.
In case of error recovery requests, this 'expires' value used
to be set to 0, so in case of a lost interrupt, such a recovery
request would never expire and block the device.

To prevent this kind of problem, all recovery requests need to
have an expires value > 0 as well. If not specified otherwise,
this should be the same expires value as for the original request.

Bugzilla
--------
BZ 768890
https://bugzilla.redhat.com/show_bug.cgi?id=768890

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.3
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=1afcfd5948ff27cdbc6d91e9f3cdbdd7f3b1e566

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>