add network priority cgroup infrastructure
Patchwork-id: 43749
O-Subject: [RHEL 6.3 PATCH 1/3] net: add network priority cgroup infrastructure
	(v2)
Bugzilla: 772974
RH-Acked-by: David S. Miller <davem@redhat.com>

From: Neil Horman <nhorman@tuxdriver.com>

This patch adds in the infrastructure code to create the network priority
cgroup.  The cgroup, in addition to the standard processes file creates two
control files:

1) prioidx - This is a read-only file that exports the index of this cgroup.
This is a value that is both arbitrary and unique to a cgroup in this subsystem,
and is used to index the per-device priority map

2) priomap - This is a writeable file.  On read it reports a table of 2-tuples
<name:priority> where name is the name of a network interface and priority is
indicates the priority assigned to frames egresessing on the named interface and
originating from a pid in this cgroup

This cgroup allows for skb priority to be set prior to a root qdisc getting
selected. This is benenficial for DCB enabled systems, in that it allows for any
application to use dcb configured priorities so without application modification

(cherry picked from commit 5bc1421e34ecfe0bd4b26dc3232b7d5e25179144)

Change notes:
v2) Added some recently posted patches upstream that fix a few off by one
errors, and one found by Stanislaw.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>