efivars: String functions
Patchwork-id: 43806
O-Subject: [RHEL6.3 PATCH 20/31] efivars: String functions
Bugzilla: 696383
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Matthew Garrett <mjg@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=696383

Commit a2940908391f3cee72e38769b30e829b22742b5b
Author: Mike Waychison <mikew@google.com>
Date: Thu, 21 Jul 2011 16:57:57 -0400

Fix the string functions in the efivars driver to be called utf16_*
instead of utf8_* as the encoding is utf16, not utf8.

As well, rename utf16_strlen to utf16_strnlen as it takes a maxlength
argument and the name should be consistent with the standard C function
names.  utf16_strlen is still provided for convenience in a subsequent
patch.

Signed-off-by: Mike Waychison <mikew@google.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>