efivars: Use string functions in pstore_write
Patchwork-id: 43808
O-Subject: [RHEL6.3 PATCH 22/31] efivars: Use string functions in pstore_write
Bugzilla: 696383
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Matthew Garrett <mjg@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=696383

Commit c475594d838c5c872e734f693a700df8c01b39d4
Author: Mike Waychison <mikew@google.com>
Date: Thu, 21 Jul 2011 16:57:59 -0400

Instead of open-coding the string operations for comparing the prefix of
the variable names, use the provided utf16_* string functions.

This patch also changes the calls to efi.set_variable to
efivars->ops->set_variable so that the right function gets called in the
case of gsmi (which doesn't have a valid efi structure).

As well, make sure that we only consider variables with the right vendor
string.

Signed-off-by: Mike Waychison <mikew@google.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>