pstore: change mutex locking to spin_locks
Patchwork-id: 43802
O-Subject: [RHEL6.3 PATCH 26/31] pstore: change mutex locking to spin_locks
Bugzilla: 696383
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Matthew Garrett <mjg@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=696383

Commit abd4d5587be911f63592537284dad78766d97d62
Author: Don Zickus <dzickus@redhat.com>
Date: Fri, 12 Aug 2011 10:54:51 -0700

pstore was using mutex locking to protect read/write access to the
backend plug-ins.  This causes problems when pstore is executed in
an NMI context through panic() -> kmsg_dump().

This patch changes the mutex to a spin_lock_irqsave then also checks to
see if we are in an NMI context.  If we are in an NMI and can't get the
lock, just print a message stating that and blow by the locking.

All this is probably a hack around the bigger locking problem but it
solves my current situation of trying to sleep in an NMI context.

Tested by loading the lkdtm module and executing a HARDLOCKUP which
will cause the machine to panic inside the nmi handler.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Acked-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>