pstore: pass reason to backend write callback
Patchwork-id: 43804
O-Subject: [RHEL6.3 PATCH 29/31] pstore: pass reason to backend write callback
Bugzilla: 696383
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Matthew Garrett <mjg@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=696383

Commit 3d6d8d20ec4fd3b256632edb373a9c504724b8a9
Author: Kees Cook <keescook@chromium.org>
Date: Thu, 17 Nov 2011 13:13:29 -0800

This allows a backend to filter on the dmesg reason as well as the pstore
reason. When ramoops is switched to pstore, this is needed since it has
no interest in storing non-crash dmesg details.

Drop pstore_write() as it has no users, and handling the "reason" here
has no obviously correct value.

Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>