ecryptfs: fix regression that prevents umount of underlying filesystem
Patchwork-id: 43868
O-Subject: [PATCH RHEL6.3] ecryptfs: fix regression that prevents umount of
	underlying filesystem
Bugzilla: 766554
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

This is a fix for bug 766554.

A regression was introduced in commit ea065665ad509fa9eca54fb6cfa037816d507e42
that prevents the underlying filesystem from being unmounted once an ecryptfs
instance was mounted/unmounted over it. An additional path_lookup() is performed to
determine whether the underlying fs is ecryptfs, but the path reference is
not freed in the success return path. Change the goto to ensure we always
put the path reference on return.

-- 1.7.7.6

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>