NFSv4: Save the owner/group name string when doing open
Patchwork-id: 43982
O-Subject: [RHEL6.3 PATCH] BZ#739797: NFSv4: Save the owner/group name string
	when doing open
Bugzilla: 739797
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

This patch fixes a problem with idmapping. With an asynchronous RPC
call, we don't want to upcall to the idmapper since that could block
rpciod and cause a deadlock. Unfortunately, that means that the
ownership ends up being set to nfsnobody for a time, even though the
NFS code marks it as needing revalidation.

In the meantime, the VFS can end up picking up and using the incorrect
ownership, which can cause certain operations to fail. This patch fixes
that by copying off the owner strings and then doing the actual
idmapping later and in normal process context.

Tested using the provided reproducer and it seems to fix the problem.
Upstream discussion has also mentioned that this bug can cause coredumps
on NFS to fail along with other problems with building software on
NFS mounts.

Original patch description from Trond follows:

---------------------[snip]--------------------

...so that we can do the uid/gid mapping outside the asynchronous RPC
context.
This fixes a bug in the current NFSv4 atomic open code where the client
isn't able to determine what the true uid/gid fields of the file are,
(because the asynchronous nature of the OPEN call denies it the ability
to do an upcall) and so fills them with default values, marking the
inode as needing revalidation.
Unfortunately, in some cases, the VFS will do some additional sanity
checks on the file, and may override the server's decision to allow
the open because it sees the wrong owner/group fields.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>