update rhg-brew target to specify the special build root
Patchwork-id: 44091
O-Subject: Re: [RHEL6 build]: update rhg-brew target to specify the special
	build root (v2)
Bugzilla:
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Ok, version 2, updated with Dons notes.  I was gonig to try clean up where we
attach the --scratch, but I think this close to 6.3 its best if we just do it
like this, and look at the flags cleanup separately.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>