vfs: automount should ignore LOOKUP_FOLLOW
Patchwork-id: 44154
O-Subject: [RHEL-6.3 PATCH 1/4] vfs: automount should ignore LOOKUP_FOLLOW
Bugzilla: 745775
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Miklos Szeredi <mszeredi@suse.cz>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=745775

Upstream commit: 0ec26fd0698285b31248e34bf1abb022c00f23d6

- edit: Ian Kent <ikent@redhat.com>
This upstream change reverts the semantic change to stat(2) and
related system calls which maintains the existing semantics of
autofs mount triggering.

I still believe the new behaviour is the better way to go but
I expect upstream to merge this and continue discussion on it
later.

edit-2: Ian Kent <ikent@redhat.com>
I'm coming around to thinking that we will need to live with this
behaviour, mainly due to bug 745775
- end edit-2

Indeed, the bug quoted below is the only known problem so far
and a change to resolve was merged upstream some time ago.

But for RHEL-5 it is probably the right choice though and keeping
it is the sensible thing to do, regardless of what happens upstream.
- edit end

Prior to 2.6.38 automount would not trigger on either stat(2) or
lstat(2) on the automount point.

After 2.6.38, with the introduction of the ->d_automount()
infrastructure, stat(2) and others would start triggering automount
while lstat(2), etc. still would not.  This is a regression and a
userspace ABI change.

Problem originally reported here:

  http://thread.gmane.org/gmane.linux.kernel.autofs/6098

It appears that there was an attempt at fixing various userspace tools
to not trigger the automount.  But since the stat system call is
rather common it is impossible to "fix" all userspace.

This patch reverts the original behavior, which is to not trigger on
stat(2) and other symlink following syscalls.

Reported-by: Leonardo Chiquitto <leonardo.lists@gmail.com>
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
CC: David Howells <dhowells@redhat.com>
CC: Ian Kent <raven@themaw.net>
CC: stable@kernel.org
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>