af_iucv: performance improvements for new HS transport
Patchwork-id: 44158
O-Subject: [RHEL6.3 PATCH 1/2] [s390x]af_iucv: performance improvements for new
	HS transport
Bugzilla: 786996
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Description
-----------
The close and send socket operation takes a long time.

For the close call, a socket state change is missing and, hence,
add the state change to IUCV_CLOSED in the afiucv_hs_callback_txnotify.
Also improve locking.

For the send call, the net_device handling is simplified and the
main part is moved to bind to accelerate sending.  Poll blocks
sending if message limit is reached.

In addition the usage of ancillary data is enabled with
HiperSockets transport.

Bugzilla
--------
BZ 786996
https://bugzilla.redhat.com/show_bug.cgi?id=786996

Upstream status of the patch
----------------------------
The patch has been submitted and is accepted in the net-text tree:
http://git.kernel.org/?p=linux/kernel/git/davem/net-next.git;a=commitdiff;h=7f1b0ea42a800713a3d56e1e8ca1a845e0461ca2
http://git.kernel.org/?p=linux/kernel/git/davem/net-next.git;a=commitdiff;h=800c5eb7b5eba6cb2a32738d763fd59f0fbcdde4
http://git.kernel.org/?p=linux/kernel/git/davem/net-next.git;a=commitdiff;h=42bd48e0145567acf7b3d2ae48bea765315bdd89
http://git.kernel.org/?p=linux/kernel/git/davem/net-next.git;a=commitdiff;h=c64d3f8f59367e89e83582b50bf072474ba2abff

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>