Fix bootmem allocator large bitmap
Patchwork-id: 44165
O-Subject: [RHEL6.3 PATCH 5/9] Fix bootmem allocator large bitmap
Bugzilla: 737748
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

This area of the kernel has changed substantially so the
problem does not exist upstream.

The allocator uses a large bitmap to manage early boot allocations. The map is
indexed by a page number. The default macros only support a max index of 2G
bits.  An index > 2g will set bits outside the boundaries of the array. This
can corrupt memory or worse (best?) case, panic the system.

Any system with over 8TB of memory is at risk of hitting this bug. However,
depending on the exact config and memory allocations, you may or may not
actually hit the bug.  Any change in the system configuration could result in
hitting or not hitting the bug.  Even if memory is corrupted, since most memory
is unused early in boot, there is a chance that the corruption is harmless.

The fix is trivial and has been tested on both UV1 and UV2.

(Note: this may be responsible for some of the strange panics we've seen on
RHEL with large memory configs).

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>