perf record: Add ability to record event period
Patchwork-id: 44235
O-Subject: [PATCH RHEL6.3 BZ788156 35/79] perf record: Add ability to record
	event period
Bugzilla: 788156
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 788156
https://bugzilla.redhat.com/show_bug.cgi?id=788156

upstream status
===============
commit 3e76ac78b08479e84a3eca3fb1b3066fb8230461
Author: Andrew Vagin <avagin@openvz.org>
Date:   Tue Dec 20 17:32:45 2011 +0300

description
===========
The problem is that when SAMPLE_PERIOD is not set, the kernel generates
a number of samples in proportion to an event's period. Number of these
samples may be too big and the kernel throttles all samples above a
defined limit.

E.g.: I want to trace when a process sleeps. I created a process which
sleeps for 1ms and for 4ms.  perf got 100 events in both cases.

swapper 0 [000] 1141.371830: sched_stat_sleep: comm=foo pid=1801 delay=1386750 [ns]
swapper 0 [000] 1141.369444: sched_stat_sleep: comm=foo pid=1801 delay=4499585 [ns]

In the first case a kernel want to send 4499585 events and in the second
case it wants to send 1386750 events.  perf-reports shows that process
sleeps in both places equal time.

Instead of this we can get only one sample with an attribute period. As
result we have less data transferring between kernel and user-space and
we avoid throttling of samples.

The patch "events: Don't divide events if it has field period" added a
kernel part of this functionality.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>