tg3: Check transitions to D0 power state
Patchwork-id: 44355
O-Subject: [RHEL6.3 PATCH 15/75] tg3: Check transitions to D0 power state
Bugzilla: 740052
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

Currently pci_set_power_state() does not return useful return codes for
transitions to the D0 power state.  If a device refuses to go into D0,
the PCI layer issues a warning but returns success.

Entering into D0 is a requirement for correct operation of tg3 devices
though.  If the PCI layer should be changed to return an error code for
this type of failure, the tg3 driver would be interested in catching it
and reacting to it.  This patch makes the necessary modifications.

commit: bed9829f3ec4b30d5cbec6511a11368ba5f169fb
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>