tg3: Determine PCI function number in one place
Patchwork-id: 44368
O-Subject: [RHEL6.3 PATCH 16/75] tg3: Determine PCI function number in one place
Bugzilla: 740052
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

tg3 devices will need to know exactly what function number they are so
that they can communicate their status to the other functions.  In a KVM
environment, the function number of a device presented by the kernel
might not be the true function number, so an alternative method to
determine the function number is needed.

This patch used to contain an implementation for the alternative method,
but recently we discovered a hardware bug that renders it incorrect.
While new method is not yet known, it is still useful to consolidate the
code that determines the PCI function to one location and use the
results throughout the code.

commit: 69f11c9936ba310ed3620072983da6ed95e86cc0
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>