tg3: Fix link down notify failure when EEE disabled
Patchwork-id: 44371
O-Subject: [RHEL6.3 PATCH 22/75] tg3: Fix link down notify failure when EEE
	disabled
Bugzilla: 740052
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

Occasionally, when the network cable is removed after a successful
autonegotiation, the device will not send a link down interrupt to the
driver.  This happens because of a bad interaction of an EEE
workaround.  The fix is to adjust the code so that the root cause
condition does not happen.

commit: b715ce947f51c6637e78b262501f0c4ff9d848cc
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>