md/raid1: perform bad-block tests for WriteMostly devices too
Patchwork-id: 44425
O-Subject: [RHEL6.3 PATCH 1/1] md/raid1: perform bad-block tests for WriteMostly
	devices too.
Bugzilla: 789570
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

From: NeilBrown <neilb@suse.de>

BZ#789570

We normally try to avoid reading from write-mostly devices, but when
we do we really have to check for bad blocks and be sure not to
try reading them.

With the current code, best_good_sectors might not get set and that
causes zero-length read requests to be send down which is very
confusing.

This bug was introduced in commit d2eb35acfdccbe2 and so the patch
is suitable for 3.1.x and 3.2.x

Reported-and-tested-by: Michał Mirosław <mirq-linux@rere.qmqm.pl>
Reported-and-tested-by: Art -kwaak- van Breemen <ard@telegraafnet.nl>
Signed-off-by: NeilBrown <neilb@suse.de>
Cc: stable@vger.kernel.org
(cherry picked from commit 307729c8bc5b5a41361af8af95906eee7552acb1)

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>