DM RAID: Set MD_CHANGE_DEVS on device rebuild
Patchwork-id: 44498
O-Subject: [RHEL6.3 PATCH 1 of 2] DM RAID: Set MD_CHANGE_DEVS on device rebuild
Bugzilla: 754271
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bug #754271:
 https://bugzilla.redhat.com/show_bug.cgi?id=754271

Brew Build:
 http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4029896

Testing:
Specifying 'rebuild' successfully kicks off a full rebuild of the
designated devices and commits changes to RAID superblock.

Upstream Status:
 Staged for 3.3 by agk@redhat.com

 brassow

dm-raid: 'rebuild' parameter code should be setting MD_CHANGE_DEVS

The 'rebuild' parameter is used to rebuild individual devices in an array.
(Resynchronize a RAID1 device or recalculate a parity device in higher RAID,
for example.)  The MD_CHANGE_DEVS flag must be set when this parameter is
given in order to write out the superblocks and make the change take immediate
effect.

So, we need MD_CHANGE_DEVS to be set and the 'FirstUse' flag was being set as
a special case for rebuilds in super_init_validaion.  The code that handles
new devices in super_load already sets MD_CHANGE_DEVS and 'FirstUse'.  Adding
a condition for rebuilds in super_load takes care of both flags without a need
for the special case in 'super_init_validation'.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>