DM RAID: Handle failure to read superblock
Patchwork-id: 44508
O-Subject: [RHEL6.3 PATCH 2 of 2] DM RAID: Handle failure to read superblock
Bugzilla: 759567
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bug #759567:
 https://bugzilla.redhat.com/show_bug.cgi?id=759567

Brew Build:
 http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4029896

Testing:
 If a superblock is unable to be read (due to a failure), the device is
properly marked as failed and the array is activated.  (Previously, the
array would fail to activate at all.)

Upstream Status:
 Staged for 3.3 by agk@redhat.com

 brassow

DM RAID: Fix inability to handle failed devices during start-up

The code fails to create a RAID array if any of the superblocks cannot be read.
However, if the userspace program passing in the device-mapper table spots the
failure and passes in the sentinel values ('- -') for the failed array position,
then everything is handled properly.  This functional gap should not exist.

This patch makes it so that if a superblock cannot be read, the array position's
fields are initialized as though '- -' were passed in via the CTR table.  That
is, the device is failed and the position should not be used; but if there is
sufficient redundancy, the array should still be activated.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>