md: Fix sysctl check failure
Patchwork-id: 44559
O-Subject: [RHEL6 PATCH] Fix sysctl check failure
Bugzilla: 714021
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=714021

This bug was seen while booting kernel-debug in a kvm
guest.  There were some cherry-picked changes to fix
Bug 688725 that led to this.  The actual RHEL6 commit
is cdbb8035b81a11922610966da3ffbd0f0c66cf30 .

>From the BZ comments:

"When md_init calls register_sysctl_table with this raid_root_table,
eventually it gets checked by sysctl_check_dir.  This will look up
the namespace in the root_table and compare the procname and ctl_name
between it and the table being registerd, and if they don't match,
it causes the failure seen in this bz."

Complete details are in Comment 7 of the bug

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>