bridge: Reset IPCB when entering IP stack
Patchwork-id: 44672
O-Subject: [RHEL6 PATCH] bridge: Reset IPCB when entering IP stack
Bugzilla: 717407
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

RHEL6 Bugzilla #717407

This is a backport of upstream patch

commit 6b1e960fdbd75dcd9bcc3ba5ff8898ff1ad30b6e
Author: Herbert Xu <herbert@gondor.apana.org.au>
Date:   Fri Mar 18 05:27:28 2011 +0000

    bridge: Reset IPCB when entering IP stack on NF_FORWARD

with an extra hunk added as the corresponding code was removed
upstream.

It fixes a crash in netfilter when used together with bridging.

bridge: Reset IPCB when entering IP stack

Whenever we enter the IP stack proper from bridge netfilter we
need to ensure that the skb is in a form the IP stack expects
it to be in.

The entry point on NF_FORWARD did not meet the requirements of
the IP stack, therefore leading to potential crashes/panics.

This patch fixes the problem.

RHEL6 still has a hook on LOCAL_OUT (which no longer exists
upstream) that also needs this treatment.  I have updated the
patch accordingly.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

Thanks,

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>