Fix deadlock in hid_reset when Dell iDRAC is reset
Patchwork-id: 44865
O-Subject: [RHEL6.3 PATCH] usb: Fix deadlock in hid_reset when Dell iDRAC is
	reset
Bugzilla: 782374
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Description:
When Dell iDRAC is reset, the iDRAC's USB keyboard/mouse device stops responding but
is not actually disconnected.  This causes usbhid to hid hid_io_error(), and
you get a chain of calls like...

hid_reset()
 usb_reset_device()
  usb_reset_and_verify_device()
   usb_ep0_reinit()
    usb_disble_endpoint()
     usb_hcd_disable_endpoint()
      ehci_endpoint_disable()

Along the way, as a result of an error/timeout with a USB transaction,
ehci_clear_tt_buffer() calls usb_hub_clear_tt_buffer() (to clear a failed
transaction out of the transaction translator in the hub), which schedules
hub_tt_work() to be run (using keventd), and then sets qh->clearing_tt=1 so
that nobody will mess with that qh until the TT is cleared.

But run_workqueue() never happens for the keventd workqueue on that CPU, so
hub_tt_work() never gets run.  And qh->clearing_tt never gets changed back to
0.

This causes ehci_endpoint_disable() to get stuck in a loop waiting for
qh->clearing_tt to go to 0.

Part of the problem is hid_reset() is itself running on keventd.  So when that thread gets a timeout trying to talk to the
HID device, it schedules clear_work (to run hub_tt_work) to run, and then gets stuck in ehci_endpoint_disable waiting for it to run.

However, clear_work never gets run because the workqueue for that CPU is still waiting for hid_reset to finish.  The workqueue code underwent a major update in 2.6.36 (See upstream section for more info!)

A much less invasive patch for earlier kernels is to just schedule clear_work on khubd if the usb code needs to clear the TT and it sees that it is already running on keventd.  Khubd isn't used by default because it can get blocked by device enumeration sometimes, but I think it should be ok for a backup for unusual cases like this just to prevent deadlock.

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=782374

Upstream:
This is an RH only patch because workqueue implementation would need to be rehauled to 2.6.36 levels.

The work queue commits between 2.6.35 and 2.6.36 changed considerably because of the workqueue concurrency management feature
See..
http://lwn.net/Articles/403891/
and
http://lwn.net/Articles/355700/

This means that to be upstream-like the workqueue implementation would need a big change that may be counter effective. Instead this approach is contained for RHEL6 for the usecase where hid_reset and usb_clear_tt_buffer are both running under keventd. Also, since khubd does not run very often and is typically used for rescue operations such as these it is logical.

Testing:
1) Dell tested the patch on the affected system and a few other systems and passed testing.
Note: The deadlock does not happen on all systems with iDRAC as it is timing dependent. Observed and tested on R310/T310 and tested on a few other systems for regression

2) https://beaker.engineering.redhat.com/jobs/192788

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=4035564

Kabi
No symbols were affected.

Thanks,
Shyam Iyer
Dell Onsite Engineer

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>